store: move RocskDB and TestDB sturcts to their separate modules
This helps separate things which relates to the Database interface
from code that is RocksDB and TestDB specific.
