+++ core/store/src/config.rs
@@ -1,8 +1,6 @@
 use near_primitives::shard_layout::ShardUId;
 use near_primitives::version::DbVersion;
 
-use crate::db::Mode;
-
 const STORE_PATH: &str = "data";
 
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
@@ -47,6 +45,19 @@ pub struct StoreConfig {
     /// We're still experimenting with this parameter and it seems decreasing its value can improve
     /// the performance of the storage
     pub trie_cache_capacities: Vec<(ShardUId, usize)>,
+}
+
+/// Mode in which to open the storage.
+#[derive(Clone, Copy)]
+pub enum Mode {
+    ReadOnly,
+    ReadWrite,
+}
+
+impl Mode {
+    pub fn read_write(rw: bool) -> Self {
+        if rw { Self::ReadWrite } else { Self::ReadOnly }
+    }
 }
 
 impl StoreConfig {
