+++ neard/src/cli.rs
@@ -7,7 +7,6 @@ use near_o11y::{
 };
 use near_primitives::types::{Gas, NumSeats, NumShards};
 use near_state_viewer::StateViewerSubCommand;
-use near_store::db::{Mode, RocksDB};
 use std::cell::Cell;
 use std::net::SocketAddr;
 use std::path::{Path, PathBuf};
@@ -70,7 +69,7 @@ impl NeardCmd {
             NeardSubCommand::Run(cmd) => cmd.run(&home_dir, genesis_validation, runtime),
 
             NeardSubCommand::StateViewer(cmd) => {
-                let mode = if cmd.readwrite { Mode::ReadWrite } else { Mode::ReadOnly };
+                let mode = near_store::Mode::read_write(cmd.readwrite);
                 cmd.subcmd.run(&home_dir, genesis_validation, mode);
             }
 
@@ -428,7 +427,7 @@ impl RunCmd {
         });
         sys.run().unwrap();
         info!(target: "neard", "Waiting for RocksDB to gracefully shutdown");
-        RocksDB::block_until_all_instances_are_dropped();
+        near_store::db::RocksDB::block_until_all_instances_are_dropped();
     }
 }
